+++ src/main/java/com/vonage/client/numbers/ListNumbersFilter.java
@@ -18,17 +18,35 @@ package com.vonage.client.numbers;
 import com.vonage.client.QueryParamsRequest;
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.Objects;
 
+/**
+ * Filter criteria used in {@link NumbersClient#listNumbers(ListNumbersFilter)}.
+ */
 public class ListNumbersFilter implements QueryParamsRequest {
     private Integer index, size;
     private String pattern;
     private SearchPattern searchPattern;
     private Type type;
 
+    private ListNumbersFilter(Builder builder) {
+        if ((index = builder.index) != null && index < 1) {
+            throw new IllegalArgumentException("'index' must be positive.");
+        }
+        if ((size = builder.size) != null && (size < 1 || size > 100)) {
+            throw new IllegalArgumentException("'size' must be between 1 and 100.");
+        }
+        pattern = builder.pattern;
+        searchPattern = builder.searchPattern;
+        type = builder.type;
+    }
+
+    @Deprecated
     public ListNumbersFilter() {
         this(null, null, null, null);
     }
 
+    @Deprecated
     public ListNumbersFilter(
             Integer index,
             Integer size,
@@ -74,7 +92,7 @@ public class ListNumbersFilter implements QueryParamsRequest {
 
     @Override
     public Map<String, String> makeParams() {
-        LinkedHashMap<String, String> params = new LinkedHashMap<>(4);
+        LinkedHashMap<String, String> params = new LinkedHashMap<>(8);
         if (index != null) {
             params.put("index", index.toString());
         }
@@ -90,7 +108,82 @@ public class ListNumbersFilter implements QueryParamsRequest {
         return params;
     }
 
-    public static class Builder {
+    /**
+     * Entrypoint for constructing an instance of this class.
+     *
+     * @return A new Builder.
+     * @since 8.10.0
+     */
+    public static Builder builder() {
+        return new Builder();
+    }
 
+    /**
+     * Builder for setting the parameters of ListNumbersFilter.
+     *
+     * @since 8.10.0
+     */
+    public static final class Builder {
+        private Integer index, size;
+        private String pattern;
+        private SearchPattern searchPattern;
+        private Type type;
+
+        private Builder() {}
+
+        /**
+         * TODO description
+         *
+         * @param index The TODO.
+         * @return This builder.
+         */
+        public Builder index(Integer index) {
+            this.index = index;
+            return this;
+        }
+
+        /**
+         * TODO description
+         *
+         * @param size The TODO.
+         * @return This builder.
+         */
+        public Builder size(Integer size) {
+            this.size = size;
+            return this;
+        }
+
+        /**
+         * TODO description
+         *
+         * @param strategy
+         * @param pattern The TODO.
+         * @return This builder.
+         */
+        public Builder pattern(SearchPattern strategy, String pattern) {
+            this.searchPattern = Objects.requireNonNull(strategy, "Matching strategy is required.");
+            this.pattern = Objects.requireNonNull(pattern, "Pattern is required");
+            return this;
+        }
+
+        /**
+         * TODO description
+         *
+         * @param type The TODO.
+         * @return This builder.
+         */
+        public Builder type(Type type) {
+            this.type = type;
+            return this;
+        }
+
+        /**
+         * Builds the ListNumbersFilter request parameters.
+         *
+         * @return A new ListNumbersFilter instance with this builder's properties.
+         */
+        public ListNumbersFilter build() {
+            return new ListNumbersFilter(this);
+        }
     }
 }