+++ src/main/java/com/vonage/client/numbers/NumbersClient.java
@@ -55,10 +55,9 @@ public class NumbersClient {
      *
      * @return A ListNumbersResponse containing the first 10 phone numbers
      *
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      */
-    public ListNumbersResponse listNumbers() throws VonageResponseParseException, VonageClientException {
+    public ListNumbersResponse listNumbers() throws NumbersResponseException {
         return listNumbers(new ListNumbersFilter());
     }
 
@@ -69,13 +68,11 @@ public class NumbersClient {
      *
      * @return A ListNumbersResponse containing phone numbers matching the supplied filter.
      *
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      */
-    public ListNumbersResponse listNumbers(ListNumbersFilter filter) throws VonageResponseParseException, VonageClientException {
+    public ListNumbersResponse listNumbers(ListNumbersFilter filter) throws NumbersResponseException {
         return listNumbers.execute(filter);
     }
-
 
     /**
      * Search for available Vonage Virtual Numbers.
@@ -84,10 +81,9 @@ public class NumbersClient {
      *
      * @return Available Vonage Virtual Numbers.
      *
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      */
-    public SearchNumbersResponse searchNumbers(String country) throws VonageResponseParseException, VonageClientException {
+    public SearchNumbersResponse searchNumbers(String country) throws NumbersResponseException {
         return searchNumbers(new SearchNumbersFilter(country));
     }
 
@@ -95,11 +91,11 @@ public class NumbersClient {
      * Search for available Vonage Virtual Numbers.
      *
      * @param filter search for available Vonage Virtual Number with filters
-     * @return available Vonage Virtual Numbers
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @return The available Vonage Virtual Numbers.
+     *
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      */
-    public SearchNumbersResponse searchNumbers(SearchNumbersFilter filter) throws VonageResponseParseException, VonageClientException {
+    public SearchNumbersResponse searchNumbers(SearchNumbersFilter filter) throws NumbersResponseException {
         return searchNumbers.execute(filter);
     }
 
@@ -109,10 +105,9 @@ public class NumbersClient {
      * @param country The two character country code in ISO 3166-1 alpha-2 format.
      * @param msisdn  The phone number to be bought in E.164 format.
      *
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      */
-    public void buyNumber(String country, String msisdn) {
+    public void buyNumber(String country, String msisdn) throws NumbersResponseException {
         buyNumber(country, msisdn, null);
     }
 
@@ -125,12 +120,11 @@ public class NumbersClient {
      *                     account here. If you’d like to perform an action on your own account,
      *                     you do not need to provide this field.
      *
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      * @see #buyNumber(String, String)
      * @since 8.10.0
      */
-    public void buyNumber(String country, String msisdn, String targetApiKey) {
+    public void buyNumber(String country, String msisdn, String targetApiKey) throws NumbersResponseException {
         buyNumber.execute(new BuyCancelNumberRequest(country, msisdn, targetApiKey));
     }
 
@@ -140,10 +134,9 @@ public class NumbersClient {
      * @param country The two character country code in ISO 3166-1 alpha-2 format.
      * @param msisdn  The phone number to be cancelled in E.164 format.
      *
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      */
-    public void cancelNumber(String country, String msisdn) {
+    public void cancelNumber(String country, String msisdn) throws NumbersResponseException {
         cancelNumber(country, msisdn, null);
     }
 
@@ -156,12 +149,11 @@ public class NumbersClient {
      *                     account here. If you’d like to perform an action on your own account,
      *                     you do not need to provide this field.
      *
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      * @see #cancelNumber(String, String)
      * @since 8.10.0
      */
-    public void cancelNumber(String country, String msisdn, String targetApiKey) {
+    public void cancelNumber(String country, String msisdn, String targetApiKey) throws NumbersResponseException {
         cancelNumber.execute(new BuyCancelNumberRequest(country, msisdn, targetApiKey));
     }
 
@@ -170,10 +162,9 @@ public class NumbersClient {
      *
      * @param request Details of the updates to be made to the number association.
      *
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      */
-    public void updateNumber(UpdateNumberRequest request) throws VonageResponseParseException, VonageClientException {
+    public void updateNumber(UpdateNumberRequest request) throws NumbersResponseException {
         updateNumber.execute(request);
     }
 
@@ -184,10 +175,9 @@ public class NumbersClient {
      * @param country The country for the given msisdn.
      * @param appId   The ID for the Vonage Application to be associated with the number.
      *
-     * @throws VonageResponseParseException if the response from the API could not be parsed.
-     * @throws VonageClientException        if an error is returned by the server.
+     * @throws NumbersResponseException If the API call returned an unsuccessful (4xx or 5xx) response.
      */
-    public void linkNumber(String msisdn, String country, String appId) throws VonageResponseParseException, VonageClientException {
+    public void linkNumber(String msisdn, String country, String appId) throws NumbersResponseException {
         updateNumber(UpdateNumberRequest.builder(msisdn, country)
                 .voiceCallback(UpdateNumberRequest.CallbackType.APP, appId).build());
     }