+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added `targetApiKey` for buy & cancel number endpoints
 - Added Builder for `UpdateNumberRequest` and deprecated setters
 - Added `app_id` property to `UpdateNumberRequest` (replaces `messagesCallbackValue`)
+- Deprecated setters in Numbers API
 
 # [8.9.4] - 2024-07-25
 - Fixed UUID validation in `ConversationAction.Builder#canHear` and `canSpeak`